# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Comment:
    properties:
      body:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      postId:
        type: integer
    type: object
  domain.Post:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
  requests.CommentRequest:
    properties:
      body:
        description: Email string `json:"email"`
        type: string
      name:
        description: |-
          PostId int64  `json:"postId,omitempty"`
          Id     int64  `json:"id,omitempty"`
        type: string
    required:
    - body
    - name
    type: object
  requests.PostRequest:
    properties:
      body:
        type: string
      title:
        description: |-
          UserId int64 `json:"userId,omitempty"`
          Id     int64  `json:"id,omitempty"`
        type: string
    required:
    - body
    - title
    type: object
  requests.UserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  resources.AuthDto:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/resources.UserDto'
    type: object
  resources.CommentDto:
    properties:
      body:
        type: string
      id:
        type: integer
      name:
        type: string
      post_id:
        type: integer
    type: object
  resources.GoogleUrlDto:
    properties:
      url:
        type: string
    type: object
  resources.PostDto:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
  resources.UserDto:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for NIX_Education application.
  title: NIX_Education API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: log in user
      parameters:
      - description: User body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.UserRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.AuthDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Log in user
      tags:
      - auth
  /auth/loginGoogle:
    post:
      consumes:
      - application/json
      description: log in user with Google (1)
      produces:
      - application/json
      - text/xml
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.GoogleUrlDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Log in user with Google (1)
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: register a new user
      parameters:
      - description: User body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.UserRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.AuthDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new user
      tags:
      - auth
  /posts:
    get:
      consumes:
      - application/json
      description: get all posts
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PostDto'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Show all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: save a post
      parameters:
      - description: Post body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.PostRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.PostDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: delete post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: get post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PostDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Show a post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: update post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.PostRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.PostDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update post
      tags:
      - posts
  /posts/{postId}/comments:
    get:
      consumes:
      - application/json
      description: get all comments
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.CommentDto'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Show all comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: save a comment
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CommentRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.CommentDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new comment
      tags:
      - comments
  /posts/{postId}/comments/{id}:
    delete:
      consumes:
      - application/json
      description: delete comment by ID
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete comment
      tags:
      - comments
    get:
      consumes:
      - application/json
      description: get comment by ID
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.CommentDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Show a comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: update comment by ID
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CommentRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.CommentDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update comment
      tags:
      - comments
  /users/{id}:
    put:
      consumes:
      - application/json
      description: set password for user google acc
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.UserRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.UserDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set password for user google acc
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"